Rework a ton of core APIs!

+++ src/shifter/shifter_trrs.h
@@ -28,9 +28,9 @@ class ShifterTrrs : public Shifter
 {
 public:
     ShifterTrrs(esp_err_t *can_init_status, const char *name, bool *start_enable);
-    ShifterPosition get_shifter_position(const uint64_t now, const uint64_t expire_time_ms) override;
-    ProfileSwitchPos get_shifter_profile_switch_pos(const uint64_t now, const uint64_t expire_time_ms) override;
-    void update_shifter_position(const uint64_t now);
+    ShifterPosition get_shifter_position(const uint32_t now, const uint32_t expire_time_ms) override;
+    ProfileSwitchPos get_shifter_profile_switch_pos(const uint32_t now, const uint32_t expire_time_ms) override;
+    void update_shifter_position(const uint32_t now);
 
 private:
     const uint8_t IO_ADDR = 0x20u;
@@ -46,7 +46,7 @@ private:
     ShifterPosition last_valid_position = ShifterPosition::SignalNotAvailable;
     uint8_t i2c_rx_bytes[2] = {0, 0};
     uint8_t i2c_tx_bytes[2] = {0, 0};
-    uint64_t last_i2c_query_time = 0u;
+    uint32_t last_i2c_query_time = 0u;
     bool *_start_enable;
 };
